+++ Sources/TuistGenerator/Generator/BuildPhaseGenerator.swift
@@ -194,9 +194,7 @@ final class BuildPhaseGenerator: BuildPhaseGenerating {
         pbxTarget.buildPhases.append(sourcesBuildPhase)
 
         var buildFilesCache = Set<AbsolutePath>()
-        // Ignore DocC Swift tutorial files from `Sources`
         let sortedFiles = files
-            .filter { !fileElements.isDocCTutorialFile(path: $0.path) }
             .sorted(by: { $0.path < $1.path })
 
         var pbxBuildFiles = [PBXBuildFile]()
@@ -378,14 +376,6 @@ final class BuildPhaseGenerator: BuildPhaseGenerating {
             let pathString = buildFilePath.pathString
             let isLocalized = pathString.contains(".lproj/")
             let isLproj = buildFilePath.extension == "lproj"
-            let isWithinAssets = pathString.contains(".xcassets/") || pathString.contains(".scnassets/")
-
-            /// Assets that are part of a .xcassets or .scnassets folder
-            /// are not added individually. The whole folder is added
-            /// instead as a group.
-            if isWithinAssets {
-                return
-            }
 
             var element: (element: PBXFileElement, path: AbsolutePath)?