+++ projects/tuist/features/step_definitions/build.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require "xcodeproj"
 
 Then(/^tuist builds the project$/) do
   system(@tuist, "build", "--path", @dir)
@@ -7,6 +8,25 @@ end
 Then(/^tuist builds the scheme ([a-zA-Z\-]+) from the project$/) do |scheme|
   system(@tuist, "build", scheme, "--path", @dir)
 end
+
+Then(/^tuist builds the scheme ([a-zA-Z\-]+) from the project with device "(.+)"$/) do |scheme, device|
+    args = [
+      "-scheme", scheme,
+      "-sdk", "iphonesimulator",
+    ]
+    if @workspace_path.nil?
+      args.concat(["-project", @xcodeproj_path]) unless @xcodeproj_path.nil?
+    else
+      args.concat(["-workspace", @workspace_path]) unless @workspace_path.nil?
+    end
+
+    args.concat(["-destination", "'name=#{device}'"])
+
+    args << "build"
+
+    xcodebuild(*args)
+end
+
 
 Then(%r{^tuist builds the scheme ([a-zA-Z\-]+) from the project at ([a-zA-Z/]+)$}) do |scheme, path|
   system(@tuist, "build", scheme, "--path", File.join(@dir, path))